Separated out the source code closure bit of code
